% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{dp_disconnect}
\alias{dp_disconnect}
\title{Close a datapond connection}
\usage{
dp_disconnect(con)
}
\arguments{
\item{con}{A connection from \code{\link[=dp_connect]{dp_connect()}}.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Close a datapond connection
}
