% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{dp_info}
\alias{dp_info}
\title{Information about one database}
\usage{
dp_info(id)
}
\arguments{
\item{id}{Database id.}
}
\value{
The registry entry (a list of class \code{datapond_db}), printed in a
readable layout.
}
\description{
Information about one database
}
\examples{
\dontrun{
dp_info("eoir")
}
}
