% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dp_local_path}
\alias{dp_local_path}
\title{Path of a locally downloaded database}
\usage{
dp_local_path(id, must_exist = FALSE)
}
\arguments{
\item{id}{Database id (see \code{\link[=dp_list]{dp_list()}}).}

\item{must_exist}{Error if the file is not present?}
}
\value{
A path (character scalar).
}
\description{
Path of a locally downloaded database
}
\examples{
dp_local_path("eoir")
}
