% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapond-package.R
\docType{package}
\name{datapond-package}
\alias{datapond}
\alias{datapond-package}
\title{datapond: Query Curated 'DuckDB' Databases Built from Public Data}
\description{
Connects to the 'datapond' registry of curated 'DuckDB' databases built from public government and research data (immigration courts, campaign finance, clinical trials, Medicare, and more). Databases are attached remotely over HTTP so only the byte ranges a query touches are transferred, or downloaded once for local use. Returns standard 'DBI' connections that work with 'dbplyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datapond-db/datapond-r}
  \item \url{https://datapond-db.github.io/website/}
  \item Report bugs at \url{https://github.com/datapond-db/datapond-r/issues}
}

}
\author{
\strong{Maintainer}: Ian Nason \email{ign.nason@gmail.com}

Authors:
\itemize{
  \item Ian Nason \email{ign.nason@gmail.com}
}

}
\keyword{internal}
